business logic is separated from auth components
