+++ want_to_do/src/components/Auth/SignIn.tsx
@@ -1,17 +1,18 @@
 import React, {useEffect} from "react";
-import {useNavigate} from "react-router-dom";
 import {useSignInUser} from "../../hooks/user";
 import AuthForm from "./AuthForm";
+import useToIntroPage from "./hooks/useToIntroPage";
+import {useNavigate} from "react-router-dom";
 
 const SignIn = () => {
-
   const navigate = useNavigate();
+  const toIntroPage = useToIntroPage();
   const { mutate } = useSignInUser();
 
   useEffect(() => {
     const accessToken = localStorage.getItem("accessToken");
     if (accessToken) {
-      navigate("/");
+      navigate("/home");
     }
   }, [navigate]);
 
@@ -19,15 +20,11 @@ const SignIn = () => {
     mutate({email, password});
   }
 
-  const toIntro = () => {
-    navigate("/");
-  }
-
   return (
       <AuthForm submitLabel={"로그인"}
                 handleSubmit={handleSignIn}
                 cancelLabel={"취소"}
-                handleCancel={toIntro}
+                handleCancel={toIntroPage}
       />
   )
 }
