+++ want_to_do/src/components/Auth/hooks/useToIntroPage.ts
@@ -1,0 +1,11 @@
+import {useNavigate} from "react-router-dom";
+
+const useToIntroPage = () => {
+  const navigate = useNavigate();
+
+  return () => {
+    navigate('/');
+  }
+}
+
+export default useToIntroPage;
